# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: user Data Access API 2
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

paths:
  /dbStats:
    get:
      description: Return DB Stats
      x-swagger-router-controller: users
      operationId: dbStats

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/dbStatsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
            
  /users:
    get:
      description: Get users
      x-swagger-router-controller: users
      operationId: getUsers
      
      parameters:
        -
          name: size
          in: query
          description: Size of array
          required: true
          type: number
          format: double

      responses:
        "200":
          description: Successful response
          # Use JSON Schema format
          schema:
            $ref: "#/definitions/userSchema"

    post:
      description: Post user (INSERT)
      x-swagger-router-controller: users
      operationId: createUser

      parameters:
        - in: body
          name: body
          description: Add a user
          required: false
          schema:
            $ref: "#/definitions/userSchema"
            
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/{id}: 
    get:             
       description: Returns a user based on a passed id
       x-swagger-router-controller: users
       operationId: getSingleUser
       
       parameters: 
         -
           name: id
           in:  path
           description: ID of user to fetch
           required: true
           type: string

       responses:
         "201":
            description: Null response
         default:
           description: unexpected error
           schema:
             $ref: "#/definitions/ErrorResponse"            

    put:
      description: Put user (UPDATE)
      x-swagger-router-controller: users
      operationId: updateUser
       
      parameters:
        - 
          name: id
          in:  path
          description: Update a user
          required: true
          type: string
          
        - name: body
          in: body
          description: User to update
          required: false
          schema:
            $ref: "#/definitions/userSchema"
  
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
        "405":
          description: Validation exception   
          
    delete: 
      description: deletes a single user
      x-swagger-router-controller: users
      operationId: deleteUser

      parameters: 
        - 
          name: id
          in: path
          description: ID of user to delete
          required: true
          type: integer
          format: int64
      
      responses: 
        "204":
          description: user deleted
        default: 
          description: unexpected error
          schema: 
            $ref: "#/definitions/ErrorResponse"         

  /swagger:
    x-swagger-pipe: swagger_raw
    
# complex objects have schema definitions
definitions:

  userSchema:
  
    required:
    
      - userId
      - userName
      - password
      - fName
      - lName
      
    properties:
    
      userId:
        type: integer
      userName:
        type: string
      password:
        type: string
      fName: 
        type: string
      lName:
        type: string

  dbStatsResponse:
    required:
      - message
    properties:
      message:
        type: string
        
  ErrorResponse:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string  